refactor: move migrations to server folder
